perf(treesitter)!: incremental and batched invalidation
Incremental invalidation:
When setting included regions, match them with the existing regions so
that they can be parse incrementally. It matches regions by
"similarity", because some changes of regions cannot be precisely
tracked by `_edit()`. Existing regions that do not match are
invalidated.

Batched invalidation:
Eagerly invalidating non-matching regions for each `parse()` is not
efficient if multiple windows display different ranges of the same
buffer. To improve efficiency, batch the invalidation.

Breaking change:
The indices of `parser:trees()` behave differently when new regions are
added manually because they are first added and then the stale regions
are discarded. So the existing uses of `trees[1]` may break. Use
`next(trees())` instead.