feat(decor): allow nvim_set_decoration_provider to skip initial redraw
This enables the initial call to nvim_set_decoration_provider to skip a
redraw. It has no effect on subsequent redraws. This is useful when a
decoration provider is first installed, but a redraw may not yet be
wanted (for example, on startup we want to prevent an immediate redraw
to avoid clearing the intro text).