build: introduce Justfile
A justfile is basically equivalent to a Makefile, except that it works
cross-platform. This helps mitigate the problem of neovim being hard to
build on non-unix. It requires the executable
[just](https://github.com/casey/just) which is readily available on
pretty much all platforms:

The Makefile is not removed as this is supposed to be a trial period,
but it may make sense to remove the Makefile in the future in case the
Justfile can replace all functionality of the makefile.

Running a just target is the same as running a make target:

```
just # builds neovim
just functionaltest # run functionaltest
```

To list all available targets with a summary of what they do:

```
just --list
```

which outputs

```
Available recipes:
    deps           # Build dependencies
    functionaltest # Run functionaltest
    nvim           # Build neovim
```

The documentaiton can be found at https://just.systems/man/en/.