vim-patch:103f1dfb7df3
runtime(less): reset readonly setting when disabling less mode

While at it, also do the unmap only if a specific key has been mapped.
There are some keys that are only selectively mapped and it would cause
an error if we are trying to unmap such a key (e.g. z when your
foldmethod is not manual).

fixes: vim/vim#14040

https://github.com/vim/vim/commit/103f1dfb7df350650a5d7caadb0364bd79e9d25b

Co-authored-by: Christian Brabandt <cb@256bit.org>