vim-patch:315cd1fbcbba
runtime(spec): Recognize SourceLicense tag name in RPM spec syntax (#14046)

rpm-4.19.0 added a new SourceLicense tag. It is used at the same place
as License tag.

This patch adds the new tag name into a Vim syntax file to be
highligted the same way as the License tag. Note that it has to be
defined in the syntax file before Source\d* regexp. Otherwise it's not
recognized by Vim.

https://github.com/vim/vim/commit/315cd1fbcbba7c44ec8743f03645bfcaef58bd55

Co-authored-by: Petr Pisar <ppisar@redhat.com>