perf(treesitter): bound `has-ancestor?` for highlighter
Problem:
`has-ancestor?` predicate can take long when the match is deep in the
tree, e.g., long `else if` chain in C. This may slow down highlighting
and thus redrawing significantly.

Solution:
Add option for limiting `has-ancestor?`'s search length, and use it for
highlighter.
The limit `15` is arbitrary, but seems to be good enough:
* Existing use of `has-ancestor?` in highlight queries are quite local.
  For example in cpp, it searches for a call expression above a
  quantified identifier to highlight the function name.
* Scrolling is reponsive enough in the long `else if` chain example.