Add maximum pattern index from prior tree to subpriority
When the highlighter has multiple highlight states (e.g. when using
injections), the pattern index in subsequent trees may be lower than the
pattern index in the "base" tree, so the extmarks for later trees are
hidden by the extmarks from earlier trees. Without subpriority this is
not an issue since extmarks appear in the (reverse) order in which they
are applied, but we _must_ use subpriorities when using iter_matches
because we cannot rely on the ordering returned by iter_matches (which
may return query patterns out of order).

To fix this, we track the maximum pattern index in each tree. Then in
each subsequent tree, we offset the pattern index used for the
subpriority by the sum of all of the maximum pattern indices from prior
trees.

For example, consider a buffer that uses the Lua highlighter and has C
injections. The highlight state will have the Lua highlighter as the
first tree and the C highlighter as the second tree. The highlighter
will first iterate through the Lua highlight query and apply extmarks.
As we iterate, we keep track of the largest pattern index (call this X).
When we then iterate through the C highlight query, a pattern might have
index Y which is less than X. If we don't adjust the subpriority, the
Lua highlight will appear over the C highlight and the injection will
not be visible. If we adjust the subpriority of the C highlight to X +
Y, then the injection is guaranteed to be visible over the Lua
highlight.