vim-patch:5cd86c6cff94
runtime(vim): Update base-syntax, improve number matching (vim/vim#14175)

- Limit blob literals to an even number of hexadecimal digits and
  correctly located dots.
- Match octal numbers.

The current version unsuccessfully attempts to match a leading '-' as
part of the float literal.  It's actually parsed as part of the literal
for all numbers but the syntax file hasn't matched it like that for a
long time and highlights negative numbers as UNARY-MINUS NUMBER.  This
will be fixed when better expression matching is implemented.

https://github.com/vim/vim/commit/5cd86c6cff94256ed2db872c46b57da259a648ac

Co-authored-by: dkearns <dougkearns@gmail.com>